requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.1
billiard==3.6.4.0
boto3==1.26.63
botocore==1.29.63
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-silk==5.0.3
django-storages==1.13.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.27.0
flower==1.2.0
gprof2dot==2022.7.29
gunicorn==20.1.0
humanize==4.5.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
Pillow==9.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.1
redis==4.4.2
referencing==0.32.0
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.13.2
s3transfer==0.6.0
simplejson==3.18.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
